requirements.txt
bottle>=0.12.9
Jinja2>=2.8
python-dotenv>=0.5.1
shapely>=1.5.17
pyproj>=1.9.5.1
gevent
six>=1.10.0
pytz>=2016.4
requests
uwsgi
boto3
pytesseract>=0.1.6
Pillow>=4.0.0
GDAL==2.1.0
affine
